refactor: extract stage update logic to comply with 200-line limit
- Created useBedStageUpdate hook (189 lines)
- Reduced BedDashboardClient from 242 to 73 lines
- Removed console.log debug statement
- All functionality preserved, no breaking changes
- Fixes file size compliance for final checks
